src/analytics/log_analytics.py
"""
Advanced Log Analytics Module for Enterprise SIEM
Handles log collection, aggregation, correlation, and real-time analysis
"""
import logging
import json
from typing import Dict, List, Optional, Union, Generator
from dataclasses import dataclass
from datetime import datetime
import uuid
import apache_beam as beam
from apache_beam.options.pipeline_options import PipelineOptions
from elasticsearch import Elasticsearch
from kafka import KafkaConsumer, KafkaProducer
import numpy as np
from sklearn.ensemble import IsolationForest
from ..core.utils import encrypt_data, decrypt_data
from ..core.database import Database

@dataclass
class LogEvent:
    """Represents a log event."""
    id: str
    timestamp: datetime
    source: str
    source_type: str
    event_type: str
    severity: str
    message: str
    raw_data: Dict
    metadata: Dict
    tags: List[str]

@dataclass
class CorrelatedEvent:
    """Represents a correlated event pattern."""
    id: str
    pattern_name: str
    events: List[LogEvent]
    correlation_type: str
    severity: str
    confidence: float
    detection_time: datetime
    context: Dict

class LogAnalytics:
    """Advanced log analytics system with real-time processing capabilities."""
    
    def __init__(self, config_path: str = None):
        self.logger = logging.getLogger(__name__)
        self.db = Database()
        self.es = Elasticsearch()
        self._initialize_analytics()
        
    def _initialize_analytics(self) -> None:
        """Initialize analytics components."""
        try:
            # Initialize Kafka
            self.consumer = KafkaConsumer(
                'logs',
                bootstrap_servers=['localhost:9092'],
                auto_offset_reset='latest',
                enable_auto_commit=True,
                group_id='log_analytics'
            )
            
            self.producer = KafkaProducer(
                bootstrap_servers=['localhost:9092'],
                value_serializer=lambda x: json.dumps(x).encode('utf-8')
            )
            
            # Initialize ML models
            self._initialize_ml_models()
            
            # Initialize Apache Beam pipeline
            self._initialize_beam_pipeline()
            
        except Exception as e:
            self.logger.error(f"Error initializing analytics: {str(e)}")
            
    def collect_logs(self, source: str) -> Generator[LogEvent, None, None]:
        """Collect logs from various sources."""
        try:
            for message in self.consumer:
                log_data = json.loads(message.value)
                
                # Parse and normalize log
                log_event = self._parse_log(log_data)
                
                # Enrich log with context
                enriched_log = self._enrich_log(log_event)
                
                # Store log
                self._store_log(enriched_log)
                
                yield enriched_log
                
        except Exception as e:
            self.logger.error(f"Error collecting logs: {str(e)}")
            
    def process_logs_realtime(self) -> None:
        """Process logs in real-time using Apache Beam."""
        try:
            options = PipelineOptions()
            
            with beam.Pipeline(options=options) as pipeline:
                # Read logs
                logs = (pipeline 
                       | 'ReadLogs' >> beam.io.ReadFromKafka(
                           consumer_config={'bootstrap_servers': 'localhost:9092'},
                           topics=['logs']
                       ))
                
                # Parse logs
                parsed_logs = (logs 
                             | 'ParseLogs' >> beam.Map(self._parse_log))
                
                # Enrich logs
                enriched_logs = (parsed_logs 
                               | 'EnrichLogs' >> beam.Map(self._enrich_log))
                
                # Detect anomalies
                anomalies = (enriched_logs 
                           | 'DetectAnomalies' >> beam.Map(self._detect_anomalies))
                
                # Correlate events
                correlated = (enriched_logs 
                            | 'CorrelateEvents' >> beam.Map(self._correlate_events))
                
                # Generate alerts
                alerts = (correlated 
                         | 'GenerateAlerts' >> beam.Map(self._generate_alerts))
                
                # Store results
                alerts | 'StoreAlerts' >> beam.Map(self._store_alert)
                
        except Exception as e:
            self.logger.error(f"Error processing logs: {str(e)}")
            
    def correlate_events(self, events: List[LogEvent]) -> List[CorrelatedEvent]:
        """Correlate events to identify patterns."""
        try:
            correlated_events = []
            
            # Apply correlation rules
            for rule in self._get_correlation_rules():
                matches = self._apply_correlation_rule(events, rule)
                if matches:
                    correlated_event = CorrelatedEvent(
                        id=str(uuid.uuid4()),
                        pattern_name=rule['name'],
                        events=matches,
                        correlation_type=rule['type'],
                        severity=rule['severity'],
                        confidence=self._calculate_confidence(matches, rule),
                        detection_time=datetime.now(),
                        context=self._build_correlation_context(matches, rule)
                    )
                    correlated_events.append(correlated_event)
                    
            return correlated_events
            
        except Exception as e:
            self.logger.error(f"Error correlating events: {str(e)}")
            return []
            
    def detect_anomalies(self, events: List[LogEvent]) -> List[Dict]:
        """Detect anomalies in log events."""
        try:
            anomalies = []
            
            # Prepare data for anomaly detection
            features = self._extract_features(events)
            
            # Apply isolation forest
            iso_forest = IsolationForest(contamination=0.1, random_state=42)
            predictions = iso_forest.fit_predict(features)
            
            # Process anomalies
            for idx, pred in enumerate(predictions):
                if pred == -1:  # Anomaly
                    anomaly = {
                        'event': events[idx],
                        'score': iso_forest.score_samples([features[idx]])[0],
                        'detection_time': datetime.now(),
                        'features': features[idx].tolist(),
                        'context': self._build_anomaly_context(events[idx])
                    }
                    anomalies.append(anomaly)
                    
            return anomalies
            
        except Exception as e:
            self.logger.error(f"Error detecting anomalies: {str(e)}")
            return []
            
    def analyze_patterns(self, events: List[LogEvent]) -> Dict:
        """Analyze patterns in log events."""
        try:
            analysis = {
                'timestamp': datetime.now(),
                'temporal_patterns': self._analyze_temporal_patterns(events),
                'behavioral_patterns': self._analyze_behavioral_patterns(events),
                'statistical_analysis': self._perform_statistical_analysis(events),
                'trend_analysis': self._analyze_trends(events),
                'recommendations': self._generate_pattern_recommendations(events)
            }
            
            return analysis
            
        except Exception as e:
            self.logger.error(f"Error analyzing patterns: {str(e)}")
            return {}
            
    def generate_analytics_report(self) -> Dict:
        """Generate comprehensive analytics report."""
        try:
            report = {
                'timestamp': datetime.now(),
                'summary': self._generate_summary(),
                'event_statistics': self._calculate_event_statistics(),
                'anomaly_analysis': self._analyze_anomalies(),
                'correlation_analysis': self._analyze_correlations(),
                'pattern_analysis': self._analyze_patterns(),
                'recommendations': self._generate_recommendations()
            }
            
            return report
            
        except Exception as e:
            self.logger.error(f"Error generating analytics report: {str(e)}")
            return {}
            
    def get_analytics_dashboard(self) -> Dict:
        """Get log analytics dashboard data."""
        try:
            dashboard = {
                'real_time_metrics': self._get_realtime_metrics(),
                'event_distribution': self._get_event_distribution(),
                'anomaly_metrics': self._get_anomaly_metrics(),
                'correlation_metrics': self._get_correlation_metrics(),
                'pattern_metrics': self._get_pattern_metrics(),
                'system_health': self._get_system_health()
            }
            
            return dashboard
            
        except Exception as e:
            self.logger.error(f"Error getting analytics dashboard: {str(e)}")
            return {}
