requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
sqlalchemy==2.0.23
# Using SQLite for development
# psycopg2-binary==2.9.9
elasticsearch==8.11.0
redis==5.0.1
kafka-python==2.0.2

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.14.0
torch==2.1.1
pandas==2.1.3
numpy==1.26.2

# Additional ML Dependencies
transformers==4.35.2

# Security
cryptography==41.0.5
pyOpenSSL==23.3.0
certifi==2023.11.17

# Testing Dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
coverage==7.3.2
pytest-mock==3.12.0
responses==0.24.1

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Network Scanning & Security
python-nmap==0.7.1
requests==2.31.0
paramiko==3.3.1
scapy==2.5.0
pypcap==1.3.0
netaddr==0.9.0

# Compliance & Reporting
pyyaml==6.0.1
jinja2==3.1.2
reportlab==4.0.7
markdown==3.5.1
python-docx==1.0.1

# Email & Notifications
slackclient==2.9.4
aiosmtplib==2.0.2
python-socketio==5.10.0

# Database
sqlite3-api==2.0.1
alembic==1.12.1
sqlmodel==0.0.11

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7
statsd==4.0.1

# Agent Dependencies
watchdog==3.0.0

# STIX/TAXII Integration
stix2==3.0.1
stix2-patterns==2.0.0
taxii2-client==2.3.0

# System Information
distro==1.8.0
psutil==5.9.6
wmi==1.5.1; platform_system == "Windows"
